refactor: move reuse interface in port pkg
